pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qzbot"
version = "0.14"
authors = [
    {name = "nm123"}
]
description = "No"
requires-python = ">=3.6"
dependencies = [
    "python-socketio",
    "requests",
    "pyjwt"
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["qzbot*"]
exclude = ["tests*"]
